[IC] Improved error model
The new model defaults all errors to be internal. The idea is for user
facing code to "promote" specific errors it expects to blame it on the
user. In other words, if an error occures that the user facing code
doesn't expect, it is considered an internal error.

Also updating some old name styles (removing "Type" from the end of
protocols)